Update Kubernetes configurations to include MariaDB deployment
This commit restructures the Kubernetes configuration files. It removes old db-config and db-secret files but introduces a new ingress, several ConfigMaps, secrets, issuer, and a statefulset for MariaDB deployment. It updates the root password and user credentials and consolidates them under MariaDB. Also, it modifies the README file with updated database connection credentials.