Add new services and deployments to Kubernetes configurations
New deployments and services have been added to the Kubernetes configuration files for the backend and frontend. The changes include the addition of services for enrollment, timetable, study-plan, account, and deployments for account-service, enrollment-service, timetable-service, and two frontends. The configurations for these new additions include detailed specifications such as container images, env variables, and port settings.