Add backup persistent volume in Kubernetes
A new yaml file for a persistent volume (PV) named 'backup-pv' has been added. This PV has a storage capacity of 1Gi, follows the Filesystem volume mode, and is accessible in ReadWriteOnce mode. Persistence of the data is guaranteed through the 'Retain' reclaim policy and data is stored in "/mnt/data" on the host file system.